+++ simulator/oneshotimporter/importer_test.go
@@ -1,4 +1,4 @@
-package clusterresourceimporter
+package oneshotimporter
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 	"github.com/golang/mock/gomock"
 	"golang.org/x/xerrors"
 
-	m "sigs.k8s.io/kube-scheduler-simulator/simulator/clusterresourceimporter/mock_clusterresourceimporter"
+	m "sigs.k8s.io/kube-scheduler-simulator/simulator/oneshotimporter/mock_clusterresourceimporter"
 	"sigs.k8s.io/kube-scheduler-simulator/simulator/snapshot"
 )
