+++ simulator/server/di/di.go
@@ -10,7 +10,7 @@ import (
 	restclient "k8s.io/client-go/rest"
 	configv1 "k8s.io/kube-scheduler/config/v1"
 
-	"sigs.k8s.io/kube-scheduler-simulator/simulator/clusterresourceimporter"
+	"sigs.k8s.io/kube-scheduler-simulator/simulator/oneshotimporter"
 	"sigs.k8s.io/kube-scheduler-simulator/simulator/reset"
 	"sigs.k8s.io/kube-scheduler-simulator/simulator/resourcewatcher"
 	"sigs.k8s.io/kube-scheduler-simulator/simulator/scheduler"
@@ -19,11 +19,11 @@ import (
 
 // Container saves and provides dependencies.
 type Container struct {
-	schedulerService             SchedulerService
-	snapshotService              SnapshotService
-	resetService                 ResetService
-	importClusterResourceService ImportClusterResourceService
-	resourceWatcherService       ResourceWatcherService
+	schedulerService               SchedulerService
+	snapshotService                SnapshotService
+	resetService                   ResetService
+	oneshotClusterResourceImporter OneShotClusterResourceImporter
+	resourceWatcherService         ResourceWatcherService
 }
 
 // NewDIContainer initializes Container.
@@ -52,7 +52,7 @@ func NewDIContainer(
 	c.snapshotService = snapshotSvc
 	if externalImportEnabled {
 		extSnapshotSvc := snapshot.NewService(externalClient, c.schedulerService)
-		c.importClusterResourceService = clusterresourceimporter.NewService(snapshotSvc, extSnapshotSvc)
+		c.oneshotClusterResourceImporter = oneshotimporter.NewService(snapshotSvc, extSnapshotSvc)
 	}
 	c.resourceWatcherService = resourcewatcher.NewService(client)
 
@@ -74,10 +74,10 @@ func (c *Container) ResetService() ResetService {
 	return c.resetService
 }
 
-// ImportClusterResourceService returns ImportClusterResourceService.
+// OneshotClusterResourceImporter returns OneshotClusterResourceImporter.
 // Note: this service will return nil when `externalImportEnabled` is false.
-func (c *Container) ImportClusterResourceService() ImportClusterResourceService {
-	return c.importClusterResourceService
+func (c *Container) OneshotClusterResourceImporter() OneShotClusterResourceImporter {
+	return c.oneshotClusterResourceImporter
 }
 
 // ResourceWatcherService returns ResourceWatcherService.
